requirements.txt
ipykernel
jupyter
langchain
langchain-anthropic
langchain-community
langchain-huggingface
langchain-openai
nltk
numpy
pandas
python-dotenv
rich
rich-cli
scikit-learn
spacy
textual
torch
torchvision
transformers